+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 * :sparkles: Zipx.kt zipCompressFilesTo method add Overload method
 * :sparkles: Premisex.kt requireNotNull and checkNotNull method add Overload method
 * :sparkles: Mapx add builder, forEach, forEachIndexed, map, mapIndexed method
+* :fire: EMPTY_MAP from Collectionx move to Mapx
 
 ## 0.6.5
 * :hammer: String orContains and andContains change to extension method